requirements.txt
streamlit==1.29.0
langchain>=0.0.217
openai>=1.2
google-generativeai==0.3.2


